directory-client: ripped presence apart a bit
