+++ sfw-provider/src/provider/mod.rs
@@ -2,7 +2,7 @@ use crate::provider::client_handling::{ClientProcessingData, ClientRequestProces
 use crate::provider::mix_handling::{MixPacketProcessor, MixProcessingData};
 use crate::provider::storage::ClientStorage;
 use crypto::identity::{DummyMixIdentityPrivateKey, DummyMixIdentityPublicKey};
-use directory_client::presence::MixProviderClient;
+use directory_client::presence::providers::MixProviderClient;
 use futures::io::Error;
 use futures::lock::Mutex as FMutex;
 use log::*;
